Lock, rep and segment decoding finished.
Assembling the output of sim86 for listing 42 with nasm now yields the same binary as the reference.

+++ sim86/sim86/instruction.h
@@ -37,7 +37,7 @@ struct Register {
 struct EffectiveAddress {
 	register_index register1;
 	register_index register2;
-	i16 displacement;
+	i32 displacement;
 	u32 segment;
 	bool explicit_segment;
 };
